fix: price to use BigDecimal

+++ store/src/main/java/com/meli/store/model/Product.java
@@ -8,6 +8,7 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -24,7 +25,7 @@ public class Product {
     private String name;
     private String sku;
     private String category;
-    private Double price;
+    private BigDecimal price;
     @OneToMany(mappedBy = "product", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<StoreInventory> inventories;
     @CreationTimestamp
@@ -32,7 +33,7 @@ public class Product {
     @UpdateTimestamp
     private LocalDateTime updatedAt;
 
-    public Product(Long id, String name, String sku, String category, Double price) {
+    public Product(Long id, String name, String sku, String category, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.sku = sku;
